requirements.txt
# Web framework
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-multipart>=0.0.9

# VoIP
twilio>=9.3.0

# AI
google-generativeai>=0.8.0

# Google APIs
google-auth>=2.35.0
google-api-python-client>=2.149.0
gspread>=6.1.0

# Email
sendgrid>=6.11.0

# Scheduling
apscheduler>=3.10.0

# Config & utilities
pydantic-settings>=2.5.0
python-dotenv>=1.0.0
httpx>=0.27.0
jinja2>=3.1.0
